fs: always sync metadata in __generic_file_fsync
If there were previously both metadata and data writeback errors when
fsync is called, then it will currently take two calls to fsync() to
clear them on some filesystems.

The problem is in __generic_file_fsync, which won't try to write back
the metadata if the data flush fails. Fix this by always attempting to
write out the metadata, even when a flush of the data reports an error.

Signed-off-by: Jeff Layton <jlayton@redhat.com>