clk: samsung: s3c: Mark expected switch fall-through
In preparation to enabling -Wimplicit-fallthrough, mark switch cases
where we are expecting to fall through.

Notice that in this particular case I replaced the comma wiht a dash,
which is what GCC is expecting to find when there are other comments
in the same line in which "fall through" appears.

Signed-off-by: Gustavo A. R. Silva <gustavo@embeddedor.com>