mmc: sdhci-cadence: fix logically and structurally dead code
Currently, the code block inside the for loop will never execute
more than once, because the function returns inmediately after
the first iteration, hence the execution of the code at the second
iteration is structurally dead and, code at line 281: return 0; is
never reached.

Based on the code comments, it seems that the actual intention is
to execute the code inside the for loop twice instead of once.

So, fix this issue by removing the return statement inside the for
loop and replace the "return 0" with "return ret".

Addresses-Coverity-ID: 1468009 ("Logically dead code")
Addresses-Coverity-ID: 1468002 ("Structurally dead code")
Signed-off-by: Gustavo A. R. Silva <gustavo@embeddedor.com>